Make CustomCatalogComponent compile.
Renamed the class files (they were called customcatalogscomponent 
(removed "s" at the end of "catalog").

The CustomCatalogComponent now encapsulates a single custom catalog.  We 
create a SkyComposite in SkyMapComposite that contains all of the user's 
custom catalogs, each in its own CustomCatalogComponent.

TODO: need functions to add/remove a custom catalog: 
SkyMapComposite::addCustomCatalog(filename) and 
SkyMapComposite::removeCustomCatalog(name)

Also converted the DeepSkyObject draw functions to use QPointF instead 
of QPoint.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=479087

+++ kstars/deepskyobject.h
@@ -172,11 +172,11 @@ public:
 
 /**Draw the object's symbol on the map
 	*/
-	void drawSymbol( QPainter &psky, int x, int y, double PositionAngle, double zoom, double scale=1.0 );
+	void drawSymbol( QPainter &psky, float x, float y, double PositionAngle, double zoom, double scale=1.0 );
 
 /**Draw the Object's image on the map
 	*/
-	void drawImage( QPainter &psky, int x, int y, double PositionAngle, double zoom, double scale=1.0 );
+	void drawImage( QPainter &psky, float x, float y, double PositionAngle, double zoom, double scale=1.0 );
 
 /**Show Deep-sky object popup menu.  Overloaded from virtual 
 	*SkyObject::showPopupMenu()
